Add filtering

+++ 1.src/midterm~1-2/src-code-new/controllers/commentController.go
@@ -11,7 +11,7 @@ import (
 func CreateComment(c *gin.Context) {
 	var body struct {
 		Body   string
-		Rating int
+		Rating float32
 	}
 	params := c.Params
